pyproject.toml
[project]
name = "insurance-claim-assistant"
version = "0.1.0"
description = "AI-powered insurance claim assistant using CrewAI multi-agent system"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.17.0",
    "langchain-mistralai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "streamlit>=1.38.0",
    "python-dotenv>=1.0.0",
    "pdfplumber>=0.11.0",
    "pypdf>=4.0.0",
    "Pillow>=10.0.0",
    "litellm>=1.75.3",
    "fpdf2>=2.8.5",
    "python-docx>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = false
